# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Customer Service Agent API
  version: 0.1.0
servers:
  - url: http://localhost:3001/api/v1
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        '200':
          description: Service health
  /conversations:
    post:
      operationId: createConversation
      responses:
        '201':
          description: Anonymous conversation and one-time access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateConversationResponse'
  /conversations/{conversationId}/messages:
    post:
      operationId: sendMessage
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
            example: Bearer one-time-conversation-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Persisted user message and deterministic knowledge or safe fallback response
        '400':
          description: Invalid message
        '401':
          description: Invalid conversation credentials
  /conversations/{conversationId}/handoff-requests:
    post:
      operationId: requestHandoff
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
            example: Bearer one-time-conversation-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reasonCode]
              properties:
                reasonCode: { type: string, enum: [customer_requested] }
      responses:
        '201':
          description: Idempotent customer handoff request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoffRequestResponse'
        '401':
          description: Invalid conversation credentials
    get:
      operationId: getHandoffStatus
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current handoff request or null
        '401':
          description: Invalid conversation credentials
  /staff/handoff-requests:
    get:
      operationId: listStaffHandoffRequests
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [requested, claimed, closed]
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
            example: Staff test-only-token
      responses:
        '200':
          description: Authorized staff handoff queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffHandoffListResponse'
        '401':
          description: Missing or invalid staff identity
        '403':
          description: Insufficient staff permission
  /staff/handoff-requests/{requestId}:
    get:
      operationId: getStaffHandoffRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Authorized staff handoff request with recent messages
        '401':
          description: Missing or invalid staff identity
        '403':
          description: Insufficient staff permission
        '404':
          description: Handoff request not found
  /staff/handoff-requests/{requestId}/claim:
    post:
      operationId: claimStaffHandoffRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Claimed handoff request, idempotently
        '401':
          description: Missing or invalid staff identity
        '403':
          description: Insufficient staff permission
        '409':
          description: Handoff state conflict
  /staff/handoff-requests/{requestId}/close:
    post:
      operationId: closeStaffHandoffRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Closed handoff request, idempotently
        '401':
          description: Missing or invalid staff identity
        '403':
          description: Insufficient staff permission
        '409':
          description: Handoff state conflict
components:
  schemas:
    CreateConversationResponse:
      type: object
      required: [conversationId, accessToken, status, createdAt]
      properties:
        conversationId: { type: string, format: uuid }
        accessToken: { type: string }
        status: { type: string, enum: [active, closed] }
        createdAt: { type: string, format: date-time }
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
    Citation:
      type: object
      required: [id, title, uri]
      properties:
        id: { type: string }
        title: { type: string }
        uri: { type: string }
        version: { type: string }
        locator: { type: string }
        contentSha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    SendMessageResponse:
      type: object
      required: [conversationId, messages, agentMode, responseType, citations, handoffRecommended, handoffStatus]
      properties:
        conversationId: { type: string, format: uuid }
        messages: { type: array, items: { type: object } }
        agentMode: { type: string, enum: [mock, deterministic_knowledge, handoff] }
        responseType: { type: string, enum: [safe_unavailable, knowledge_answer, handoff_recommended, handoff_requested, handoff_pending, mock_fallback] }
        citations: { type: array, items: { $ref: '#/components/schemas/Citation' } }
        handoffRecommended: { type: boolean }
        handoffStatus: { type: string, nullable: true, enum: [requested, claimed, closed] }
        handoffRequestId: { type: string, format: uuid, nullable: true }
        assistantMessageId: { type: string, format: uuid, nullable: true }
    HandoffRequestResponse:
      type: object
      required: [conversationId, requestId, status, reasonCode, requestedAt, updatedAt, idempotent]
      properties:
        conversationId: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        status: { type: string, enum: [requested, claimed, closed] }
        reasonCode: { type: string, enum: [customer_requested] }
        requestedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        idempotent: { type: boolean }
        claimedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
    HandoffActionResponse:
      type: object
      required: [requestId, status, idempotent]
      properties:
        requestId: { type: string, format: uuid }
        status: { type: string, enum: [requested, claimed, closed] }
        idempotent: { type: boolean }
    StaffHandoffListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StaffHandoffRequest'
    StaffHandoffRequest:
      type: object
      required: [requestId, conversationId, status, reasonCode, requestedAt, updatedAt, claimedBy, claimedAt, closedAt, recentMessages]
      properties:
        requestId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        status: { type: string, enum: [requested, claimed, closed] }
        reasonCode: { type: string }
        requestedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        claimedBy: { type: string, nullable: true }
        claimedAt: { type: string, format: date-time, nullable: true }
        closedAt: { type: string, format: date-time, nullable: true }
        recentMessages: { type: array, items: { type: object } }